"Remove Lombok dependency and replace with traditional Java methods"
Deleted the Lombok dependency in pom.xml and removed relevant annotations in the UserEntity and UserDto classes. Lombok was also removed from the project lombok.config. This was done to make the code more accessible to developers who are not familiar with Lombok. In its place, conventional getter and setter methods were implemented. Moreover, parameters in the validatePassword method in UserService class were finalized to avoid mutation.